+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/RandomPlaylistParametersPanel.java
@@ -106,7 +106,8 @@ final class RandomPlaylistParametersPanel extends JPanel {
 	}
 
 	private FilterList<Entity> createGenreList(EntityConnectionProvider connectionProvider) {
-		return Components.list(FilterListModel.builder()
+		return Components.list()
+						.model(FilterListModel.builder()
 										.items(allGenres(connectionProvider))
 										.build())
 						.selectedItems()